Add moved app into the api folder
