+++ test/test_member_show.py
@@ -1,4 +1,4 @@
-from app import app
+from api.app import app
 from classes.member_show import MemberShow
 
 def test_initialize():
@@ -42,7 +42,7 @@ def test_initialize():
               "bills_cosponsored": 395,
               "missed_votes_pct": 0.69,
               "votes_with_party_pct": 98.61,
-              "votes_against_party_pct": 1.28 
+              "votes_against_party_pct": 1.28
           }]
   }
